Handle Scan for []uint8 numeric columns
